move kms to separate kms scanner test
