+++ enumeration/remote/google_kms_scanner_test.go
@@ -1,0 +1,132 @@
+package remote
+
+import (
+	"testing"
+
+	"github.com/snyk/driftctl/enumeration"
+	"github.com/snyk/driftctl/enumeration/remote/alerts"
+	"github.com/snyk/driftctl/enumeration/remote/cache"
+	"github.com/snyk/driftctl/enumeration/remote/common"
+	remoteerr "github.com/snyk/driftctl/enumeration/remote/error"
+	"github.com/snyk/driftctl/enumeration/remote/google"
+	"github.com/snyk/driftctl/enumeration/remote/google/repository"
+	"github.com/snyk/driftctl/enumeration/terraform"
+
+	"github.com/snyk/driftctl/enumeration/resource"
+	"github.com/snyk/driftctl/mocks"
+
+	testgoogle "github.com/snyk/driftctl/test/google"
+
+	terraform2 "github.com/snyk/driftctl/test/terraform"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	assetpb "google.golang.org/genproto/googleapis/cloud/asset/v1"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+func TestGoogleKmsCryptoKey(t *testing.T) {
+	cases := []struct {
+		test             string
+		assertExpected   func(t *testing.T, got []*resource.Resource)
+		response         []*assetpb.Asset
+		responseErr      error
+		setupAlerterMock func(alerter *mocks.AlerterInterface)
+		wantErr          error
+	}{
+		{
+			test:     "no kms crypto keys",
+			response: []*assetpb.Asset{},
+			assertExpected: func(t *testing.T, got []*resource.Resource) {
+				assert.Len(t, got, 0)
+			},
+		},
+		{
+			test: "multiple kms crypto keys",
+			assertExpected: func(t *testing.T, got []*resource.Resource) {
+				assert.Len(t, got, 2)
+				assert.Equal(t, "projects/playground-bruno/locations/global/keyRings/keyring-example/cryptoKeys/foo", got[0].ResourceId())
+				assert.Equal(t, "google_kms_crypto_key", got[0].ResourceType())
+
+				assert.Equal(t, "projects/playground-bruno/locations/global/keyRings/keyring-example/cryptoKeys/bar", got[1].ResourceId())
+				assert.Equal(t, "google_kms_crypto_key", got[1].ResourceType())
+			},
+			response: []*assetpb.Asset{
+				{
+					AssetType: "cloudkms.googleapis.com/CryptoKey",
+					Name:      "//compute.googleapis.com/projects/playground-bruno/locations/global/keyRings/keyring-example/cryptoKeys/foo",
+				},
+				{
+					AssetType: "cloudkms.googleapis.com/CryptoKey",
+					Name:      "//compute.googleapis.com/projects/playground-bruno/locations/global/keyRings/keyring-example/cryptoKeys/bar",
+				},
+			},
+		},
+		{
+			test: "cannot list kms crypto keys",
+			assertExpected: func(t *testing.T, got []*resource.Resource) {
+				assert.Len(t, got, 0)
+			},
+			responseErr: status.Error(codes.PermissionDenied, "The caller does not have permission"),
+			setupAlerterMock: func(alerter *mocks.AlerterInterface) {
+				alerter.On(
+					"SendAlert",
+					"google_kms_crypto_key",
+					alerts.NewRemoteAccessDeniedAlert(
+						common.RemoteGoogleTerraform,
+						remoteerr.NewResourceListingError(
+							status.Error(codes.PermissionDenied, "The caller does not have permission"),
+							"google_kms_crypto_key",
+						),
+						alerts.EnumerationPhase,
+					),
+				).Once()
+			},
+		},
+	}
+
+	factory := terraform.NewTerraformResourceFactory()
+
+	for _, c := range cases {
+		t.Run(c.test, func(tt *testing.T) {
+			scanOptions := ScannerOptions{}
+			providerLibrary := terraform.NewProviderLibrary()
+			remoteLibrary := common.NewRemoteLibrary()
+
+			// Initialize mocks
+			alerter := &mocks.AlerterInterface{}
+			if c.setupAlerterMock != nil {
+				c.setupAlerterMock(alerter)
+			}
+
+			assetClient, err := testgoogle.NewFakeAssertServerWithList(c.response, c.responseErr)
+			if err != nil {
+				tt.Fatal(err)
+			}
+
+			realProvider, err := terraform2.InitTestGoogleProvider(providerLibrary, "3.78.0")
+			if err != nil {
+				tt.Fatal(err)
+			}
+
+			repo := repository.NewAssetRepository(assetClient, realProvider.GetConfig(), cache.New(0))
+
+			remoteLibrary.AddEnumerator(google.NewGoogleComputeForwardingRuleEnumerator(repo, factory))
+
+			testFilter := &enumeration.MockFilter{}
+			testFilter.On("IsTypeIgnored", mock.Anything).Return(false)
+
+			s := NewScanner(remoteLibrary, alerter, scanOptions, testFilter)
+			got, err := s.Resources()
+			assert.Equal(tt, err, c.wantErr)
+			if err != nil {
+				return
+			}
+			alerter.AssertExpectations(tt)
+			testFilter.AssertExpectations(tt)
+			if c.assertExpected != nil {
+				c.assertExpected(t, got)
+			}
+		})
+	}
+}
